fix: Rename folder to baby
