+++ src/main/java/com/isppG8/infantem/infantem/baby/BabyService.java
@@ -1,4 +1,4 @@
-package com.isppG8.infantem.infantem.bebe;
+package com.isppG8.infantem.infantem.baby;
 
 import java.util.List;
